requirements.txt
discord.py[voice]==2.6.4
yt-dlp
PyNaCl
